refactor: Move File types to @chainlit/components
