+++ libs/components/src/types/file.ts
@@ -1,0 +1,13 @@
+export interface FileSpec {
+  accept?: string[] | Record<string, string[]>;
+  max_size_mb?: number;
+  max_files?: number;
+}
+
+export interface IFileResponse {
+  name: string;
+  path?: string;
+  size: number;
+  type: string;
+  content: ArrayBuffer;
+}
