+++ libs/components/src/types/index.ts
@@ -1,5 +1,6 @@
 export * from './action';
 export * from './element';
+export * from './file';
 export * from './Input';
 export * from './messages';
 export * from './NotificationCount';
